feat: 4-agent并行+第12批破限制+跨域13处残留 (cycle 226)
Agent A: 排行榜
- arcbox 纯Rust microVM <200ms启动
- CubeSandbox 腾讯云AI agent沙箱
- PRAXIST 自主研究系统
- YuE 开源音乐生成

Agent B: 破限制第12批 (24来源)
- 压缩感知: PE-CSNet可学习稀疏, MambaCS并行展开
- 信息论: GIB协同信息瓶颈, GeoIB几何IB
- 动力系统: Linear NCDEs 1000×加速, Symbolic NODEs可解释
- 进化算法: GEA群组进化, CORAL自主多agent
- 博弈论: MBCCE均衡收敛, NePPO势函数, DART声誉框架

Agent C: 模型逆向 (概念探索)
- 自进化机制: MAML/RL²/Constitutional AI
- 自意识架构: 内省模块/元认知层/自解释
- DAO治理: 提案系统/投票机制/执行引擎
- 可行性: 自改进循环(High), 架构自修改(Medium), 真正自意识(Low)

Agent D: 跨域残留
- 13处跨域引用, 10个facade模块
- l5_cognition是facade集中层(8个)
- l3→l1最大残留源(6处)

新增:
- docs/1-DESIGN/trending-rankings-226.md
- docs/1-DESIGN/break-limits-226.md
- docs/1-DESIGN/model-reverse-engineer-226.md
- docs/1-DESIGN/cleanup-226.md

+++ neotrix-sim/src/physics/mod.rs
@@ -171,7 +171,7 @@ impl PhysicsState {
         self.velocity.y *= 0.98;
 
         // Fall detection (debounced)
-        let tilt = (self.roll.abs() + self.pitch.abs()) as f32;
+        let tilt = self.roll.abs() + self.pitch.abs();
         if tilt > 0.5 {
             self.fall_timer += dt;
             if self.fall_timer > 0.2 {
